Handle lost-ground uncertainty in a cleaner way, move some of the vertical uncertanity in its own method.
Instead of proactively calculating speed with thisMove#touchedGroundWorkaround, let lost-ground be treated as an "after failure" check: if the player fails the horizontal prediction and a lost-ground case happened, force-set the ground flag and re-calculate (effect on performance needs be evaluated).
Same thing applies for vertical motion with the PlayerEnvelopes#isTouchDownMove thing.

More:
* At last, get rid of bunnyfly: with a stricter vertical monitoring this should not be needed anymore.

* On respawning: let the player decide if sneaking/sprinting should be reset.

* Clarify some stuff here and there.

* Unbind lost-ground from toIsValid check in AirWorkarounds (move validity is already checked in lost-ground)

* Do still register item-use events/adapters from 1.12 up to 1.17