[Broken] Sketch an inside-liquid vertical prediction.
Lots of issues still remain (this is mostly a starting point).
Namely:
1) I can't seem to find how the game handles gravity in liquid (descending; not pressing the space bar);
2) It still doesn't account for all version-dependent nuances (i.e.: 1.12 and below clients not being able to swim up to the surface), because it is a major pain in the ass to: 1. download all NMS sources 2. research and note down what changed across versions… So, what we(I?)'ll do is to just account for 1.8 clients, since that is the most used (and relevant) legacy version (at least for publishing a beta in master);
3) Vertical medium-checking order is (currently) wrong: Minecraft checks for water first, then for lava, then gliding, and finally, it falls back to air (Source: EntityLiving.java -> travel() method).
For vertical predictions, NCP check for climbing first (odd one out, because we still use a limit-based implementation for it), then gliding, and then it falls back to vdistrel, which unifies air and liquids.
I chose to trade NMS logic for code readability within NCP. Having two, almost-identical vdistrel methods (one for liquid, one for air) would be extremely ugly.
Still, it'd be best to (pedantically) follow what NMS does.
This could be solved by unifying vertical and horizontal predictions. But… Willpower, I guess? :)

The code is starting to get harder to read too, so perhaps it’s time to merge vertical and horizontal predictions to reduce redundancy anyway.